+++ pkg/postgresql/postgres_snapshot_test.go
@@ -165,7 +165,7 @@ func TestIntegrationPostgresSnapshots(t *testing.T) {
 
 			cnnstr := getCnnStr()
 
-			p, handler, err := newPostgres(t.Context(), "error", 10000, dsInfo, cnnstr, logger, backend.DataSourceInstanceSettings{})
+			p, handler, err := newPostgres(t.Context(), "error", 10000, 0, dsInfo, cnnstr, logger, backend.DataSourceInstanceSettings{})
 			require.NoError(t, err)
 
 			t.Cleanup(func() {